add a normal take damage

+++ _Game/Assets/Scripts/HealthBar.cs
@@ -7,13 +7,13 @@ public class HealthBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxHealth(int heath)
+    public void SetMaxHealth(float heath)
     {
         slider.maxValue = heath;
         slider.value = heath;
     }
     
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
     }
